Move SpanSplitEnumerator back to MemoryExtensions
